requirements.txt
certifi==2017.7.27.1
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.2
future==0.16.0
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
python-telegram-bot==8.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2
